LPD-15185 Move Pages to AplicationMenu folder
